[profiler] Optimise TickSample memory layout
- Change StateTag from int to uint16
- Make sure we don't waste any padding bytes in TickSample
- For now use full bytes for the bool values in TickSample

Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4393853
Commit-Queue: Camillo Bruni <cbruni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#86869}

+++ include/v8-profiler.h
@@ -24,7 +24,7 @@ enum class EmbedderStateTag : uint8_t;
 class HeapGraphNode;
 struct HeapStatsUpdate;
 class Object;
-enum StateTag : int;
+enum StateTag : uint16_t;
 
 using NativeObject = void*;
 using SnapshotObjectId = uint32_t;
